create clean_text.py
